Run two workflows for building in Lite and Fuse modes
