+++ .github/workflows/showcase-fuse.yml
@@ -1,10 +1,10 @@
-name: skipapp-showcase
+name: skipapp-showcase-fuse
 on:
   push:
     branches: [ main ]
     tags: "[0-9]+.[0-9]+.[0-9]+"
   schedule:
-    - cron: '30 16 * * *'
+    - cron: '30 2,6,10,14,18,22 * * *'
   workflow_dispatch:
   pull_request:
 
@@ -15,7 +15,9 @@ permissions:
 
 jobs:
   call-workflow:
-    uses: skiptools/actions/.github/workflows/skip-app.yml@v1
+    uses: skiptools/actions/.github/workflows/skip-app.yml@env-input
+    with:
+      env: SKIP_MODE=fuse
     secrets:
       # These optional secrets enable the Android app to be signed
       KEYSTORE_JKS: ${{ secrets.KEYSTORE_JKS }}
@@ -28,4 +30,3 @@ jobs:
       APPLE_MOBILEPROVISION: ${{ secrets.APPLE_MOBILEPROVISION }}
       # This secret enables the iOS app to be uploaded to the App Store
       APPLE_APPSTORE_APIKEY: ${{ secrets.APPLE_APPSTORE_APIKEY }}
-
