+++ .github/workflows/showcase-lite.yml
@@ -1,0 +1,20 @@
+name: skipapp-showcase-lite
+on:
+  push:
+    branches: [ main ]
+    tags: "[0-9]+.[0-9]+.[0-9]+"
+  schedule:
+    - cron: '30 0,4,8,12,16,20 * * *'
+  workflow_dispatch:
+  pull_request:
+
+permissions:
+  contents: write
+  id-token: write
+  attestations: write
+
+jobs:
+  call-workflow:
+    uses: skiptools/actions/.github/workflows/skip-app.yml@env-input
+    with:
+      env: SKIP_MODE=lite
